Move the location of postImageOverlayWordSchema
